Switch files over to 1 class 1 file
