+++ lib/cucumber/rails/database/deletion_strategy.rb
@@ -1,0 +1,11 @@
+module Cucumber
+  module Rails
+    module Database
+      class DeletionStrategy < Strategy
+        def before_js
+          super :deletion
+        end
+      end
+    end
+  end
+end
