+++ lib/cucumber/rails/database/strategy.rb
@@ -1,0 +1,31 @@
+module Cucumber
+  module Rails
+    module Database
+      class Strategy
+        def initialize(options = {})
+          @options = options
+        end
+
+        def before_js(strategy)
+          @original_strategy = if defined?(DatabaseCleaner::VERSION) && Gem::Version.new(DatabaseCleaner::VERSION) >= Gem::Version.new('1.8.0.beta')
+                                 DatabaseCleaner.cleaners.values.first.strategy # that feels like a nasty hack
+                               else
+                                 DatabaseCleaner.connections.first.strategy # that feels like a nasty hack
+                               end
+          DatabaseCleaner.strategy = strategy, @options
+        end
+
+        def before_non_js
+          # no-op
+        end
+
+        def after
+          return unless @original_strategy
+
+          DatabaseCleaner.strategy = @original_strategy
+          @original_strategy = nil
+        end
+      end
+    end
+  end
+end
