+++ lib/cucumber/rails/database/truncation_strategy.rb
@@ -1,0 +1,11 @@
+module Cucumber
+  module Rails
+    module Database
+      class TruncationStrategy < Strategy
+        def before_js
+          super :truncation
+        end
+      end
+    end
+  end
+end
